requirements.txt
Flask==2.3.2
face_recognition==1.3.0
cmake>=3.22.1
numpy==1.26.4
pillow==10.3.0
pandas==2.2.2
openpyxl==3.1.2
